+++ src/components/core/form/drawer-form-with-fields.tsx
@@ -1,8 +1,8 @@
 import { AnyZodSchema, useFormInputProps } from "@/utils/form-helper";
 import { InputHTMLAttributes, ReactNode } from "react";
+import { FormField } from "../../content/forms/form-field";
 import type { DrawerFormProps } from "./drawer-form";
 import { DrawerForm } from "./drawer-form";
-import { FormField } from "./form-field";
 
 export type FormEntry = {
   label: string;
