+++ src/components/core/form/drawer-form.tsx
@@ -4,10 +4,10 @@ import { useCTW } from "@/components/core/ctw-provider";
 import { AnyZodSchema } from "@/utils/form-helper";
 import { isEmpty } from "lodash";
 import { ReactNode, useState } from "react";
-import type { DrawerProps } from "../../core/drawer";
-import { Drawer } from "../../core/drawer";
-import { SaveButton } from "./save-button";
-import { ActionReturn } from "./types";
+import { SaveButton } from "../../content/forms/save-button";
+import { ActionReturn } from "../../content/forms/types";
+import type { DrawerProps } from "../drawer";
+import { Drawer } from "../drawer";
 
 export type FormErrors = Record<string, string[]>;
 type InputError = Record<string, string[]>;
